+++ server/src/main/java/org/opensearch/action/admin/indices/stats/IndicesStatsRequest.java
@@ -292,6 +292,15 @@ public class IndicesStatsRequest extends BroadcastRequest<IndicesStatsRequest> {
         return this;
     }
 
+    public boolean includeFieldLevelSegmentFileSizes() {
+        return flags.includeFieldLevelSegmentFileSizes();
+    }
+
+    public IndicesStatsRequest includeFieldLevelSegmentFileSizes(boolean includeFieldLevelSegmentFileSizes) {
+        flags.includeFieldLevelSegmentFileSizes(includeFieldLevelSegmentFileSizes);
+        return this;
+    }
+
     public IndicesStatsRequest includeUnloadedSegments(boolean includeUnloadedSegments) {
         flags.includeUnloadedSegments(includeUnloadedSegments);
         return this;