+++ server/src/test/java/org/opensearch/index/replication/IndexLevelReplicationTests.java
@@ -257,8 +257,8 @@ public class IndexLevelReplicationTests extends OpenSearchIndexLevelReplicationT
             IndexShard replica = shards.addReplica();
             shards.recoverReplica(replica);
 
-            SegmentsStats segmentsStats = replica.segmentStats(false, false);
-            SegmentsStats primarySegmentStats = shards.getPrimary().segmentStats(false, false);
+            SegmentsStats segmentsStats = replica.segmentStats(false, false, false);
+            SegmentsStats primarySegmentStats = shards.getPrimary().segmentStats(false, false, false);
             assertNotEquals(IndexRequest.UNSET_AUTO_GENERATED_TIMESTAMP, primarySegmentStats.getMaxUnsafeAutoIdTimestamp());
             assertEquals(primarySegmentStats.getMaxUnsafeAutoIdTimestamp(), segmentsStats.getMaxUnsafeAutoIdTimestamp());
             assertNotEquals(Long.MAX_VALUE, segmentsStats.getMaxUnsafeAutoIdTimestamp());